🚚 유틸 폴더 구조 재정비
