+++ Core/Util/Sources/Util/Constants/DateFormatType.swift
@@ -1,0 +1,17 @@
+//
+//  DateFormatType.swift
+//  
+//
+//  Created by Julia on 2023/07/07.
+//
+
+import Foundation
+
+public enum DateFormatType: String {
+    case yearMonthDay = "YYYY/MM/dd"
+    case hourMinute = "HH:mm"
+
+    var displayName: String {
+        return self.rawValue
+    }
+}
