--- Core/Util/Sources/Util/Scene/UIView+.swift
@@ -1,16 +1,0 @@
-//
-//  UIView+.swift
-//  
-//
-//  Created by Eddy on 2023/06/08.
-//
-
-import UIKit
-
-public extension UIView {
-    func addSubviews(_ views: UIView...) {
-        for v in views {
-            self.addSubview(v)
-        }
-    }
-}
