+++ Core/Util/Sources/Util/Traits/Scene.swift
@@ -8,12 +8,10 @@
 import UIKit
 
 public protocol Scene {
-    
     var viewController: UIViewController { get }
 }
 
 public extension Scene where Self: UIViewController {
-    
     var viewController: UIViewController {
         return self
     }
